Port sharing reproductions
